CSHARP2692: PR comments
